// src/generated/models/Cr9cd_scoringconfigsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_scoringconfigsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_scoringconfigsstatecode = keyof typeof Cr9cd_scoringconfigsstatecode;
export const Cr9cd_scoringconfigsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_scoringconfigsstatuscode = keyof typeof Cr9cd_scoringconfigsstatuscode;

export interface Cr9cd_scoringconfigsBase {
  cr9cd_is_active?: boolean;
  cr9cd_name?: string;
  cr9cd_params?: string;
  cr9cd_scoringconfigid: string;
  cr9cd_version?: number;
  cr9cd_weights?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_scoringconfigsstatecode;
  statuscode?: Cr9cd_scoringconfigsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_scoringconfigs extends Cr9cd_scoringconfigsBase {
  cr9cd_is_activename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}